feat: adiciona pacote e classe VideoManager
